requirements.txt
backports.ssl-match-hostname==3.5.0.1
requests==2.18.4
six==1.11.0
websocket-client==0.44.0
wsgiref==0.1.2
